pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenthub"
version = "0.1.0"
description = "AgentHub CLI and local tooling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml==6.0.3",
  "jsonschema==4.26.0",
  "httpx==0.28.1",
  "rich==13.9.4",
]

[project.scripts]
agenthub = "agenthub.cli:main"

[tool.setuptools]
packages = ["agenthub"]
